+++ ui/src/dialogs/detail.js
@@ -1,5 +1,3 @@
-import * as React from 'react';
-import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
@@ -8,21 +6,11 @@ import CloseIcon from '@mui/icons-material/Close';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
-import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
+import { useState } from 'react';
 
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-  '& .MuiDialogContent-root': {
-    padding: theme.spacing(2),
-  },
-  '& .MuiDialogActions-root': {
-    padding: theme.spacing(1),
-  },
-}));
-
-
-export default function FrameworksDetails({show, data}) {
-  const [open, setOpen] = React.useState(true);
+export default function TasksDetails({show, data, title}) {
+  const [open, setOpen] = useState(true);
 
   const handleClose = () => {
     setOpen(false);
@@ -47,7 +35,7 @@ export default function FrameworksDetails({show, data}) {
             {
               Object.entries(val).map(([key, value]) => (
                 <TableRow key={key}>
-                  <TableCell style={cssHeader}>{key}</TableCell>
+                  <TableCell style={cssHeader}>{key}:</TableCell>
                   <TableCell>{getValue(value)}</TableCell>
                 </TableRow>
               ))
@@ -63,14 +51,12 @@ export default function FrameworksDetails({show, data}) {
   }
 
   return (
-    <div>
-      <BootstrapDialog
+      <Dialog
         onClose={handleClose}
-        aria-labelledby="customized-dialog-title"
-        open={open}    
+        open={open}   
         maxWidth="lg"
       >
-        <DialogTitle sx={{ m: 0, p: 2 }}>
+        <DialogTitle>
           <IconButton
             aria-label="close"
             onClick={handleClose}
@@ -85,21 +71,15 @@ export default function FrameworksDetails({show, data}) {
           </IconButton>
         </DialogTitle>
         <div style={cssTitle}>
-          Show Framework: {data.id}
+         {title}: {data.id}
         </div>
         <DialogContent dividers>
           <Table sx={{ minWidth: 650 }} size="small">
-            <TableHead>
-              <TableRow>
-                <TableCell style={cssHeader}>Name</TableCell>
-                <TableCell>&nbsp;</TableCell>
-              </TableRow>
-            </TableHead>
             <TableBody>
             {
               Object.entries(data).map(([key, value]) => (
                 <TableRow key={key}>
-                  <TableCell style={cssHeader}>{key}</TableCell>
+                  <TableCell style={cssHeader}>{key}:</TableCell>
                   <TableCell>{getValue(value)}</TableCell>
                 </TableRow>
               ))
@@ -107,7 +87,6 @@ export default function FrameworksDetails({show, data}) {
             </TableBody>
           </Table>
         </DialogContent>
-      </BootstrapDialog>
-    </div>
+      </Dialog>
   );
 }