+++ changelog.md
@@ -2,7 +2,7 @@
 - Open port to the mesos agent
 - Add simple webui functionality
 - Update mini Apache Mesos to version 1.11.0-0.2.0
-- Add show details of frameworks and tasks
+- Add show details of frameworks, tasks and agents
 
 # v0.1.0
 - First inn